+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- `Class Printed\\Bundle\\Queue\\Entity\\QueueTask has no field or association named completion_percentage`
 
 ## [4.1.0] - 2017-11-24
 ### Added